refactor: expose entry-point logic for popup and content coverage
popup/index.ts and content/index.ts ran side effects at import time (the
popup queried its DOM element map at module scope; the content script
auto-ran initialize on load), so both were excluded from coverage and
their tests re-implemented logic inline without importing the real
modules — ~995 of the most side-effect-heavy lines were invisible to
the coverage gates.

Split each entry into a few-line shim (unchanged manifest/HTML paths,
import-time wiring only) and an importable logic module:

- src/content/bootstrap.ts: startContentScript/initialize/handleSync;
  the popup DOM element map moves inside initPopup() in
  src/popup/app.ts, queried at init instead of import
- Tests now drive the real modules end to end (fixture DOM, mocked
  storage/messaging/ui seams): 48 tests replacing the inline
  re-implementations
- Coverage excludes only the shims; bootstrap.ts enters measurement at
  93% stmts and app.ts at 97% stmts, global gates stay green
- platform-ssot fitness assertion follows the extractor routing to
  bootstrap.ts; ADR-015 documents the entry-shim pattern

dist wiring verified after build: manifest -> index.ts loader ->
content chunk, popup HTML -> popup chunk, service-worker-loader
basename unchanged.
